extract recovery events to module
